move activities hook to MemberActivities.hooks
