+++ packages/atlas/src/providers/notifications/notifications.types.ts
@@ -1,4 +1,4 @@
-import { BasicMembershipFieldsFragment, StorageDataObjectFieldsFragment } from '@/api/queries'
+import { BasicMembershipFieldsFragment } from '@/api/queries'
 
 export type NftNotificationRecord = {
   id: string
@@ -31,59 +31,3 @@ export type NotificationRecord = { read?: boolean } & (
       bidAmount: number
     } & NftNotificationRecord)
 )
-
-export type NftActivitiesRecord = {
-  id: string
-  date: Date
-  block: number
-  video: {
-    id: string
-    title: string
-    thumbnailPhoto: StorageDataObjectFieldsFragment | null
-  }
-}
-export type ActivitiesRecord =
-  | ({
-      type: 'Bid'
-      bidAmount: number
-      from: BasicMembershipFieldsFragment
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Purchase'
-      from: BasicMembershipFieldsFragment | null
-      price: number
-      to: BasicMembershipFieldsFragment | null
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Withdrawal'
-      from: BasicMembershipFieldsFragment
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Sale'
-      from: BasicMembershipFieldsFragment | null
-      price: number
-      to: BasicMembershipFieldsFragment | null
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Removal'
-      from: BasicMembershipFieldsFragment | null
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Transfer'
-      from: BasicMembershipFieldsFragment
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Listing'
-      typeName: 'EnglishAuctionStartedEvent' | 'OpenAuctionStartedEvent' | 'NftSellOrderMadeEvent'
-      from: BasicMembershipFieldsFragment | null
-      price?: number
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Mint'
-      from: BasicMembershipFieldsFragment | null
-    } & NftActivitiesRecord)
-  | ({
-      type: 'Price change'
-      from: BasicMembershipFieldsFragment | null
-      price: number
-    } & NftActivitiesRecord)
